+++ packages/terra-application-navigation/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.82.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.81.0 - (March 25, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.80.0 - (March 14, 2024)
 
 * Changed