+++ packages/terra-navigation-prompt/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.89.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.88.0 - (March 25, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.87.0 - (March 14, 2024)
 
 * Changed