+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+## 2.54.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 2.53.0 - (March 25, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 2.52.0 - (March 21, 2024)
+
 * Changed
   * Added accessibility label for navigation back button.