+++ packages/terra-pills/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.24.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.23.0 - (March 25, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.22.0 - (March 14, 2024)
 
 * Changed