+++ packages/terra-slide-panel-manager/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 5.99.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 5.98.0 - (March 29, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 5.97.0 - (March 25, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 5.96.0 - (March 19, 2024)
 
 * Changed