+++ packages/terra-tabs/CHANGELOG.md
@@ -2,7 +2,24 @@
 
 ## Unreleased
 
-* Updated
+## 7.25.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 7.24.0 - (March 29, 2024)
+
+* Fixed
+  * Fixes dropdown list overflow and reset left position value after close.
+
+## 7.23.0 - (March 25, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 7.22.0 - (March 21, 2024)
+
+* Changed
    * Position dropdown element only after it is opened to prevent additional white space getting added.
 
 ## 7.21.0 - (March 14, 2024)